Changed contextKinds as a Set of Integer (instead of a Set of Short)
git-svn-id: https://gama-platform.googlecode.com/svn/branches/GAMA_CURRENT@5746 b91ef52e-2b48-11df-9ab6-b1cdc0d4a89a

+++ msi.gama.core/src/msi/gaml/compilation/AbstractGamlAdditions.java
@@ -108,7 +108,7 @@ public abstract class AbstractGamlAdditions implements IGamlAdditions {
 		final String[] parentSymbols, final int[] parentKinds, final FacetProto[] fmd,
 		final String omissible, final ISymbolConstructor sc, final String ... names) {
 		Set<String> contextKeywords = new HashSet();
-		Set<Short> contextKinds = new HashSet();
+		Set<Integer> contextKinds = new HashSet();
 		final Map<String, FacetProto> facets = new HashMap();
 		if ( fmd != null ) {
 			for ( FacetProto f : fmd ) {
@@ -122,7 +122,7 @@ public abstract class AbstractGamlAdditions implements IGamlAdditions {
 		}
 		if ( parentKinds != null ) {
 			for ( int p : parentKinds ) {
-				contextKinds.add((short) p);
+				contextKinds.add(p);
 			}
 		}
